refactor: Move all test utils into `utils/testing/`
